Demo acceptance criteria for SHP3 userstory.
a) User is able to specify color of the fill in RGB8
b) User is able to specify no fill for their shapes
c) Shape is visible beneath a shape with no fill.

Draws two circles, the one behind has a fill specified in RGB8, the
second is above and has no fill.